[Decomp][Player] Cohesion split: extract spell system to PlayerSpell.cpp
Moves the contiguous spell learning + cooldown + talent-reset block
out of Player.cpp into a new file under src/game/Object/.

Player.cpp: 18,889 -> 17,702 lines (-1,187).
PlayerSpell.cpp: 1,270 lines, 12 methods.

Methods moved:
- addSpell, IsNeedCastPassiveLikeSpellAtLearn, learnSpell, removeSpell
- RemoveSpellCooldown, RemoveSpellCategoryCooldown,
  RemoveArenaSpellCooldowns, RemoveAllSpellCooldown
- _LoadSpellCooldowns, _SaveSpellCooldowns
- resetTalentsCost, resetTalents

No header change; no signature change; no behaviour change. Picked up
by the file(GLOB Object/*.cpp) rule in src/game/CMakeLists.txt.
